Module minecraft_assets::schemas::blockstates

Serde-(de)serializable data types for assets/<namespace>/blockstates/*.json.

Start here: BlockStates

See https://minecraft.fandom.com/wiki/Model#Block_states.

Modules

multipart

Types used to compose BlockStates::Multipart.

Structs

ModelProperties

Contains the properties of a model that is used to render all or part of a block in a particular state.

Enums

BlockStates

Block states as stored in the assets/<namespace>/blockstates directory.

Variant

A block variant.