Module minecraft_assets::schemas::models

Serde-(de)serializable data types for assets/<namespace>/models/{block,item}/*.json.

Start here: Model.

See https://minecraft.fandom.com/wiki/Model#Block_models.

Structs

Display

Specifies how a Model is displayed in different views.

Element

A single, cube-shaped element of a Model’s geometry.

ElementFace

Specifies the details of a single face in a cuboid Element.

ElementRotation

Specifies the rotation of an Element.

Model

A block or item model as stored in the assets/<namespace>/models/{block,item}/ directories.

OverrideCase

One possible case in which an item’s Model should be overridden.

Texture

The value of a texture variable in the Textures map.

Textures

Specifies the Textures of a Model.

Transform

Specifies the position, rotation, and scale at which a model is displayed.

Enums

Axis

The three possible axes in 3D space.

BlockFace

The six possible faces of a cuboid.

GuiLightMode

The two possible ways to shade a model in the UI.

PredicateValue

The value for an item tag specified in a predicate in an OverrideCase.